Adds daily log rotation for production.
Addresses Issue #125 - "Implement log rotation". Changes the production
log4js configuration to replace 1 file per log with a rotation of 1 file
per day per log.

This change is implemented with the log4js-node dateFile appender. As
mentioned in the documentation for this
(https://github.com/nomiddlename/log4js-node/wiki/Date-rolling-file-appender)
rolling over from a 31 day month to a 30 day month will result in the
31st day log file remaining stale until the 31st day occurs again (at
which point it will be overwritten).