Fixes #148 - pass csrf_token to template ctx.
logged-out.mustache compileAndRender call was not receiving the
csrf_token from the req.session. This results in bug #148 - a csrf token
error if you log out and then try to use the log in form from the
header.

This patch gives the logged-out mustache template the csrf_token to fix