Handle actions against suspended users better.
If all pending actions are against suspended users, currently we get stuck.
This change fixes that, and incidentally fixes an issue where we called
.thenResolve() on verifyCredentials, which doesn't return a Promise.

Also, remove the storeUser call in stream, which is generally unnecessary and
causes errors to occasionally be logged when two different streams see the same
user at nearly the same time (race condition).