Switch from Q.allSettled to Q.all
Q.allSettled makes it easy to silently fail, because an Error in the array of
Promises doesn't bubble up as an error in the returned Promise. Instead, the
caller has to remember to map through the returned objects checking for errors,
which is easy to forget (and which none of the callsites in Block Together do).

This caused this bug: https://github.com/jsha/blocktogether/issues/173