Immediately record unblocks on stream events.
For users with large numbers of blocks, we inhibit updateBlocks. However, this
means that when one of those users unblocks someone, it may be a long time
before that unblock is recorded. If that use then sends another @-reply during
that time, they'll get blocked again. Recording the unblock immediately allows
unblock protection to take effect (at the cost of occasionally duplicating an
unblock in the Actions list).