Rewrite update-blocks.
Start using async functions, which simplifies the code a lot.
Keep all blocked ids in memory until updating is done, rather than
writing them to the database piece-by-piece. This should reduce database
churn. This also makes it easier in the future to early-terminate block
updates for users with truly massive numbers of blocks.