Remove unblock-checking from actions.js.
It used to be the case that unblocking someone would prevent Block
Together from ever reblocking them. This was mainly useful in
conjunction with the auto-block features (new accounts and < 15
followers). Now that those features are gone due to Twitter API changes,
we can remove the complexity of checking for unblocks when executing
actions.